timing my_custom_objective by itself
moving our utils functions into test_poly.jl
