+++ src/Utils.jl
@@ -286,96 +286,4 @@ function dump_buffer(buffer::AnnotatedIOBuffer)
     return AnnotatedString(dump_buffer(buffer.io), buffer.annotations)
 end
 
-function fnFromString(s)
-    f = eval(Meta.parse("x -> " * s))
-    return x -> Base.invokelatest(f, x)
-end
-
-function eval_limit(f::Function, val::Float64)
-
-    @syms x
-    var = "x"
-
-    # if (occursin(var, eq_str))
-        try
-            f_val = f(val)
-            
-            if (f_val == NaN)
-                fn_value_x = f(x)
-                return limit(fn_value_x, x, val)
-            else
-                return f_val
-            end
-        catch e
-            if e isa DomainError
-                return 10000  # an arbitrary large number
-            else
-                rethrow(e)
-            end
-        end
-    # else
-    #     return 10000  # an arbitrary large number
-    # end
-end
-
-function get_deriv_from_fn(f::Function)
-
-    @syms x
-
-    # g = fnFromString(f)
-
-    deriv = expand_derivatives(Differential(x)(eval(f)(x)))
-    return build_function(deriv, x, expression=Val{false})
-end
-
-function eval_derivative(f::Function, val::Float64)
-    
-    @syms x
-    
-    # if (occursin(var, eq_str))
-
-        # new_eq_str = replace(eq_str, var => "x")
-
-        # if(eq_str == "x")
-        #     return NaN, 1
-        # end
-        
-        # f = fnFromString(new_eq_str)
-
-        try
-            df = get_deriv_from_fn(f)
-
-            if (occursin("x", repr(df)))
-                
-                df_val = df(val)
-                df_x = df(x)
-                if (df_val == NaN)
-                    return repr(df_x), limit(df_x, x, val)
-                else
-                    return repr(df_x), df_val
-                end
-            else
-                try
-                    return NaN, parse(Float64, @show simplify(df))
-                catch e
-                    if e isa ArgumentError
-                        return NaN, 10000
-                    else
-                        rethrow(e)
-                    end
-                end
-            end
-        catch e
-            if e isa DomainError
-                return NaN, 10000
-            else
-                rethrow(e)
-            end
-        end
-        
-    # else
-    #     return NaN, 10000  # a random large number
-    # end
-end
-
 end
